requirements.txt
audioread==2.1.8
cffi==1.13.2
Click==7.0
cycler==0.10.0
decorator==4.4.1
Flask==1.1.1
gunicorn==20.0.0
itsdangerous==1.1.0
Jinja2==2.10.3
joblib==0.14.0
kiwisolver==1.1.0
librosa==0.7.1
llvmlite==0.30.0
MarkupSafe==1.1.1
matplotlib==3.1.1
numba==0.46.0
numpy==1.17.4
pandas==0.25.3
pycparser==2.19
pyparsing==2.4.5
python-dateutil==2.8.1
pytz==2019.3
resampy==0.2.2
scikit-learn==0.21.3
scipy==1.3.2
six==1.13.0
sklearn-pandas==1.8.0
SoundFile==0.10.2
Werkzeug==0.16.0
youtube-dl==2019.11.5
